+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - Add support for prop completion for JSX V4 https://github.com/rescript-lang/rescript-vscode/pull/579
 - Add support for create interface file for JSX V4 https://github.com/rescript-lang/rescript-vscode/pull/580
+- Expand one level of type definition on hover. Dig into record/variant body.
 
 ## v1.6.0